test(mft): pin the extended $SI fields on parse_record_full, not just the helper
The existing pins assert `parse_standard_info_full` directly, but the
public API consumers actually call is `parse_record_full`, which reaches
those fields through two hops: a local `ExtendedStandardInfo` in `full.rs`
that is later moved onto `ParsedRecord::std_info`. A refactor giving
`full.rs` its own $SI read would keep the helper-level tests green while
silently zeroing the fields for every external caller.

Assert quota_charged/max_versions/version_number/class_id — plus the
pre-existing usn/security_id/owner_id, which ride the same two hops — on a
`ParsedRecord` returned by `parse_record_full`. Verified to fail when the
second hop is broken (quota_charged 0 != 65536).

The addition pushed parse/tests.rs to 839 LOC, past the 800-LOC policy
limit. Split rather than granted an exception: tests.rs becomes
tests/mod.rs (698 LOC, helpers and everything else) plus tests/std_info.rs
(152 LOC, all three $STANDARD_INFORMATION pins). Also drops the now-false
`single_call_fn` expectation on `create_file_name_value`, which the new
test makes a second caller of.
