+++ crates/uffs-mft/src/parse/tests/mod.rs
@@ -7,6 +7,8 @@ use super::*;
 use crate::frs::{Frs, ParentFrs};
 use crate::ntfs::{AttributeType, ExtendedStandardInfo, FILE_RECORD_MAGIC, NameInfo, ReparseTag};
 
+mod std_info;
+
 fn write_u16_le(buffer: &mut [u8], offset: usize, value: u16) {
     buffer[offset..offset + 2].copy_from_slice(&value.to_le_bytes());
 }
@@ -103,10 +105,6 @@ fn create_resident_attribute(attr_type: AttributeType, value: &[u8]) -> Vec<u8> 
     attr
 }
 
-#[expect(
-    clippy::single_call_fn,
-    reason = "test helper isolates FileName attribute layout for one targeted regression"
-)]
 fn create_file_name_value(parent_directory: u64, name: &str, namespace: u8) -> Vec<u8> {
     let name_utf16: Vec<u16> = name.encode_utf16().collect();
     let mut value = vec![0_u8; 66 + name_utf16.len() * 2];
@@ -213,86 +211,6 @@ fn apply_fixup_valid_record_on_unaligned_slice() {
     assert!(result, "Fixup should succeed for an unaligned record slice");
     assert_eq!(&storage[511..513], &0x1234_u16.to_le_bytes());
     assert_eq!(&storage[1023..1025], &0x5678_u16.to_le_bytes());
-}
-
-#[test]
-fn parse_standard_info_full_reads_unaligned_v30_payload() {
-    let attr_offset = 1_usize;
-    let value_offset = 24_u16;
-    let si_offset = attr_offset + usize::from(value_offset);
-    let mut data = vec![0_u8; si_offset + 72];
-    let creation_time = 116_444_736_000_000_010_i64;
-    let modification_time = 116_444_736_000_000_020_i64;
-    let mft_change_time = 116_444_736_000_000_030_i64;
-    let access_time = 116_444_736_000_000_040_i64;
-    let owner_id = 44_u32;
-    let security_id = 55_u32;
-    let usn = 66_u64;
-
-    write_u32_le(&mut data, attr_offset + 16, 72);
-    write_u16_le(&mut data, attr_offset + 20, value_offset);
-    write_i64_le(&mut data, si_offset, creation_time);
-    write_i64_le(&mut data, si_offset + 8, modification_time);
-    write_i64_le(&mut data, si_offset + 16, mft_change_time);
-    write_i64_le(&mut data, si_offset + 24, access_time);
-    write_u32_le(&mut data, si_offset + 48, owner_id);
-    write_u32_le(&mut data, si_offset + 52, security_id);
-    write_u64_le(&mut data, si_offset + 64, usn);
-
-    let mut result = ExtendedStandardInfo::default();
-    parse_standard_info_full(&data, attr_offset, &mut result);
-
-    assert_eq!(result.created, creation_time);
-    assert_eq!(result.modified, modification_time);
-    assert_eq!(result.mft_changed, mft_change_time);
-    assert_eq!(result.accessed, access_time);
-    assert_eq!(result.owner_id, owner_id);
-    assert_eq!(result.security_id, security_id);
-    assert_eq!(result.usn, usn);
-}
-
-/// The NTFS 3.0+ `$SI` fields that live past the timestamps — quota, version
-/// and class bookkeeping — must survive the parse, and must stay zero for the
-/// 36-byte NTFS 1.2 layout, which has no such fields.
-#[test]
-fn parse_standard_info_full_reads_quota_and_version_fields() {
-    let attr_offset = 0_usize;
-    let value_offset = 24_u16;
-    let si_offset = attr_offset + usize::from(value_offset);
-    let mut data = vec![0_u8; si_offset + 72];
-    let max_versions = 3_u32;
-    let version_number = 2_u32;
-    let class_id = 9_u32;
-    let quota_charged = 4_096_u64;
-
-    write_u32_le(&mut data, attr_offset + 16, 72);
-    write_u16_le(&mut data, attr_offset + 20, value_offset);
-    write_u32_le(&mut data, si_offset + 36, max_versions);
-    write_u32_le(&mut data, si_offset + 40, version_number);
-    write_u32_le(&mut data, si_offset + 44, class_id);
-    write_u64_le(&mut data, si_offset + 56, quota_charged);
-
-    let mut result = ExtendedStandardInfo::default();
-    parse_standard_info_full(&data, attr_offset, &mut result);
-
-    assert_eq!(result.max_versions, max_versions);
-    assert_eq!(result.version_number, version_number);
-    assert_eq!(result.class_id, class_id);
-    assert_eq!(result.quota_charged, quota_charged);
-
-    // NTFS 1.2: the same byte positions are past the end of the attribute
-    // value, so all four must remain zero.
-    let mut v12 = vec![0_u8; si_offset + 36];
-    write_u32_le(&mut v12, attr_offset + 16, 36);
-    write_u16_le(&mut v12, attr_offset + 20, value_offset);
-
-    let mut v12_result = ExtendedStandardInfo::default();
-    parse_standard_info_full(&v12, attr_offset, &mut v12_result);
-
-    assert_eq!(v12_result.max_versions, 0);
-    assert_eq!(v12_result.version_number, 0);
-    assert_eq!(v12_result.class_id, 0);
-    assert_eq!(v12_result.quota_charged, 0);
 }
 
 #[test]
